+++ setuptools/discovery.py
@@ -213,11 +213,13 @@ class FlatLayoutPackageFinder(PEP420PackageFinder):
     _EXCLUDE = (
         "ci",
         "bin",
+        "debian",
         "doc",
         "docs",
         "documentation",
         "manpages",
         "news",
+        "newsfragments",
         "changelog",
         "test",
         "tests",